+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-08T06:49:25.887Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-08T09:57:22.082Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")