+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR35JPZ86T7TJ39099J4QCPD"
+        "01KR3GATGA2YR05HTP31F9JVX0"
     )
     r"""The unique identifier of the data source"""