+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-08T06:49:25.887Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-08T09:57:22.082Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")