+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR35JPZC7G94KEHRXV4JK6VQ"
+        "01KR3GATGDSHDBAJ0Y06ABM55B"
     )
     r"""The unique identifier of the data source"""