+++ packages/orq-rc/RELEASES.md
@@ -7649,3 +7649,13 @@ Based on:
 - [python v4.9.0-rc.32] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.32 - packages/orq-rc
+
+## 2026-05-08 09:59:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.33] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.33 - packages/orq-rc